requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.25.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
anthropic==0.7.7
pydantic==2.5.0
python-multipart==0.0.6
PyPDF2==3.0.1
openpyxl==3.1.2
scikit-learn==1.3.2
python-docx==1.1.0